solidity_module/module/requirements.txt
web3
pytest
py-solc-x
python-dotenv